Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -6,10 +6,10 @@ int main() {
    int a;
    int b;
    int c;
-   scanf("%i",&a);
-   scanf("%i",&b);
-   scanf("%i",&c);
+   scanf("%d",&a);
+   scanf("%d",&b);
+   scanf("%d",&c);
    int sum = a+b+c;
-   float avg = sum / 3;
-   printf("Average: %.2f",avg);
+   double avg = sum / 3;
+   printf("Average: %.2lf",avg);
 }
